Make sources compile for WASM.
